Scoreboard works, fixed bugs with new windows and id length

+++ src/Models/Quiz.java
@@ -3,19 +3,19 @@ package Models;
 import java.util.ArrayList;
 
 public class Quiz {
-    private int _id;
+    private double _id;
     private String _name;
     private int _difficulty; // 1 - 5
     private ArrayList<Question> _questions;
 
-    public Quiz(int id, String name, int difficulty, ArrayList<Question> questions){
+    public Quiz(double id, String name, int difficulty, ArrayList<Question> questions){
         this._id = id;
         this._name = name;
         this._difficulty = difficulty;
         this._questions = questions;
     }
 
-    public int getId() {
+    public double getId() {
         return _id;
     }
